LPS-15583 Document library does not allow downloading files larger than 4GB
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@75351 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/webserver/WebServerServlet.java
@@ -402,14 +402,14 @@ public class WebServerServlet extends HttpServlet {
 			converted = true;
 		}
 
-		int contentLength = 0;
+		long contentLength = 0;
 
 		if (!converted) {
 			if (DLUtil.compareVersions(version, fileEntry.getVersion()) >= 0) {
-				contentLength = (int)fileEntry.getSize();
+				contentLength = fileEntry.getSize();
 			}
 			else {
-				contentLength = (int)fileVersion.getSize();
+				contentLength = fileVersion.getSize();
 			}
 		}
 		else {
